fix: rename exports and fix exports of importers
BREAKING_CHANGE: `handlers`, `resolver`, and `importers` was renamed to `builtinHandlers`, `builtinResolvers`, and `builtinImporters`.
